feat: refactor config handling
